requirements.txt
streamlit==1.42.2
streamlit-webrtc==0.47.9
opencv-python-headless==4.9.0.80  # Use headless for Docker
numpy==1.26.4
torch==2.2.0  # Compatible with CUDA 12.1
torchvision==0.17.0
ultralytics==8.2.2
pillow==10.3.0
pyyaml==6.0.1
aiortc==1.10.1
aioice==0.9.0
protobuf==4.25.3
ffmpeg-python==0.2.0
pyyaml